apps/backend/densecap/webcam/requirements.txt
Flask==0.10.1
Flask-Cors==2.1.2
Flask-RESTful==0.3.5
Jinja2==2.8
MarkupSafe==0.23
Pillow==3.2.0
Werkzeug==0.11.8
aniso8601==1.1.0
argparse==1.2.1
backports-abc==0.4
backports.ssl-match-hostname==3.5.0.1
certifi==2016.2.28
itsdangerous==0.24
python-dateutil==2.5.2
pytz==2016.3
singledispatch==3.4.0.3
six==1.10.0
tornado==4.3
wsgiref==0.1.2
